Cleaning, remove builder for robotframework 3.1
